Adds `JKParseOptionExtendedFloatingPoint` and `JKSerializeOptionExtendedFloatingPoint`, which allows C99 style hexadecimal floating point numbers.
When `JKParseOptionExtendedFloatingPoint` is enabled, JSON containing C99 style hexadecimal floating point numbers (i.e., case-insensitive `0x[0-9a-f]*\.[0-9a-f]p(+|-)?[0-9]+`) is permitted.  `strtod` is used to convert the parsed result.

When `JKSerializeOptionExtendedFloatingPoint` is enabled, floating point values are written to the JSON output using the `printf(3)` `"%a"` conversion specifier.

Currently, `(+|-)?(inf|infinity|nan)` are not permitted in the JSON to be parsed, nor are they written to the serialized JSON.

The benefit to using hexadecimal floating point number representation is that it allows for floating point values to be serialized and deserialized exactly.  That is to say, there will be no loss of precision due to inexact conversion to decimal representation.  These options can be useful for those who are using JSON to serialize and deserialize information to disk, much like binary plists, especially if you require the floating point values to be round tripped exactly.

Enabling `JKParseOptionExtendedFloatingPoint` "should" be 100% backwards compatible with strict parsing, but this has not been sufficiently tested at this point.

Enabling `JKSerializeOptionExtendedFloatingPoint` (and assuming there are floating point values to be serialized) **WILL** create JSON that is not strictly conforming JSON and will be rejected by other JSON implementations.

Another minor change is that serializing JSON without `JKSerializeOptionExtendedFloatingPoint` (i.e., standard JSON), floating point values are written to the JSON output using the `printf(3)` `"%.17g"`, changed from `"%.16g"`.  This should, hopefully, allow floating point values to be written out (via decimals) with up to single floating point (i.e., `float`) precision.